Brettski/affiliate payments

+++ src/dataSources/cloudFirestore/affiliatePromoCode.js
@@ -1,32 +1,47 @@
 import debug from 'debug';
 import * as Sentry from '@sentry/node';
+import { utility } from '@thatconference/api';
 
-const dlog = debug('that:api:garage:datasources:firebase:affiliate');
+const dlog = debug('that:api:garage:datasources:firebase:affiliatePromoCodes');
+const { entityDateForge } = utility.firestoreDateForge;
+const forgeFields = ['lastUpdatedAt', 'createdAt'];
+const promoCodeDateForge = entityDateForge({ fields: forgeFields });
 
-const collectionName = 'affiliates';
-const subCollectionName = 'events';
+const collectionName = 'affiliatePromoCodes';
+// const subCollectionName = 'events';
 
 const scrubPromoCode = ({ promoCode, isNew = false }) => {
   dlog('scrubbing promotion code');
   const scrubbedPromoCode = promoCode;
+  const now = new Date();
   if (isNew) {
-    scrubbedPromoCode.eventId = scrubbedPromoCode.id;
+    scrubbedPromoCode.createdAt = now;
+    scrubbedPromoCode.lastUpdatedAt = now;
   }
+  scrubbedPromoCode.lastUpdatedAt = now;
+
   return scrubbedPromoCode;
 };
 
+function createPromoCodeId({ affiliateId, eventId, instance = 0 }) {
+  return `${affiliateId}|${eventId}|${instance}`;
+}
+
+function isAffilateInPromoCodeId({ affiliateId, promoCodeId }) {
+  return promoCodeId.toLowerCase().startsWith(affiliateId.toLowerCase());
+}
+
 const affiliatePromoCode = dbInstance => {
   dlog('instance created');
-  Sentry.setTag('app location', 'affiliate store');
+  Sentry.setTag('app location', 'affiliate promoCode store');
 
-  const affiliateCollection = dbInstance.collection(collectionName);
+  const affiliatePromoCodeCollection = dbInstance.collection(collectionName);
 
-  function get({ affiliateId, promoCodeId }) {
-    dlog('get called: %s, %s', affiliateId, promoCodeId);
-    return affiliateCollection
-      .doc(affiliateId)
-      .collection(subCollectionName)
-      .doc(promoCodeId)
+  function get(affiliatePromoCodeId) {
+    dlog('get called for id %s', affiliatePromoCodeId);
+
+    return affiliatePromoCodeCollection
+      .doc(affiliatePromoCodeId)
       .get()
       .then(docSnap => {
         let result = null;
@@ -35,22 +50,26 @@ const affiliatePromoCode = dbInstance => {
             id: docSnap.id,
             ...docSnap.data(),
           };
+          result = promoCodeDateForge(result);
         }
+
         return result;
       });
   }
 
   function getAllAffiliatePromoCodes(affiliateId) {
     dlog('getAllAffiliatePromoCodes %s', affiliateId);
-    return affiliateCollection
-      .doc(affiliateId)
-      .collection(subCollectionName)
+    return affiliatePromoCodeCollection
+      .where('affiliateId', '==', affiliateId)
       .get()
       .then(querySnap =>
-        querySnap.docs.map(doc => ({
-          id: doc.id,
-          ...doc.data(),
-        })),
+        querySnap.docs.map(doc => {
+          const r = {
+            id: doc.id,
+            ...doc.data(),
+          };
+          return promoCodeDateForge(r);
+        }),
       );
   }
 
@@ -60,48 +79,86 @@ const affiliatePromoCode = dbInstance => {
     promoCodeType,
   }) {
     dlog('getAffiliatePromoCodeForEvent %s, %s', affiliateId, eventId);
-    return affiliateCollection
-      .doc(affiliateId)
-      .collection(subCollectionName)
-      .doc(eventId)
-      .get()
-      .then(docSnap => {
-        let result = null;
-        if (docSnap.exists) {
-          result = {
-            id: docSnap.id,
-            ...docSnap.data(),
-          };
-          if (promoCodeType && promoCodeType !== result.promoCodeType) {
-            result = null;
-          }
-        }
+    let query = affiliatePromoCodeCollection
+      .where('affiliatId', '==', affiliateId)
+      .where('eventId', '==', eventId);
+    if (promoCodeType) {
+      query = query.where('promoCodetype', '==', promoCodeType);
+    }
 
-        return result;
-      });
+    return query.get().then(querySnap => {
+      if (querySnap.size > 1) {
+        const err = new Error(
+          `multiple promo codes found for affiliate ${affiliateId} and event ${eventId}`,
+        );
+        Sentry.withScope(scope => {
+          scope.setTags({
+            function: 'affiliatePromoCode.getAffilatePromoCodeForEvent',
+            affiliateId,
+            eventId,
+          });
+          scope.setLevel('error');
+          scope.setContext(
+            'ids discovered',
+            querySnap.docs.map(q => q.id),
+          );
+          Sentry.captureException(err);
+        });
+        throw err;
+      }
+      const [doc] = querySnap.docs;
+      let result = null;
+      if (doc) {
+        result = {
+          id: doc.id,
+          ...doc.data(),
+        };
+        result = promoCodeDateForge(result);
+      }
+
+      return result;
+    });
+  }
+
+  function findAllPromoCodesForEvent(eventId) {
+    dlog('findAllPromoCodesForEvent %s', eventId);
+    return affiliatePromoCodeCollection
+      .where('eventId', '==', eventId)
+      .get()
+      .then(querySnap =>
+        querySnap.docs.map(doc => {
+          const r = {
+            id: doc.id,
+            ...doc.data(),
+          };
+          return promoCodeDateForge(r);
+        }),
+      );
   }
 
   function create({ affiliateId, promotionCode: newPromoCode }) {
     dlog('create %o for %s', newPromoCode, affiliateId);
-    if (!newPromoCode.id)
-      throw new Error(`id is required to create promotion code`);
+    if (!newPromoCode.eventId)
+      throw new Error(`eventId is required to create promotion code`);
     const cleanPromoCode = scrubPromoCode({
       promoCode: newPromoCode,
       isNew: true,
     });
-    const promoCodeId = cleanPromoCode.id;
-    delete cleanPromoCode.id;
-    return affiliateCollection
-      .doc(affiliateId)
-      .collection(subCollectionName)
-      .doc(promoCodeId)
+    cleanPromoCode.affiliateId = affiliateId;
+    const affiliatePromoCodeId = createPromoCodeId({
+      affiliateId,
+      eventId: cleanPromoCode.eventId,
+    });
+
+    return affiliatePromoCodeCollection
+      .doc(affiliatePromoCodeId)
       .create(cleanPromoCode)
-      .then(() => get({ affiliateId, promoCodeId }));
+      .then(() => get(affiliatePromoCodeId));
   }
 
   function update({ affiliateId, promoCodeId, promotionCode: upPromoCode }) {
     dlog(
-      'update promoCode %s for affiliate %s, %o',
+      'update event %s for affiliate %s, %o',
       promoCodeId,
       affiliateId,
       upPromoCode,
@@ -109,21 +166,24 @@ const affiliatePromoCode = dbInstance => {
     if (!affiliateId)
       throw new Error('affiliateId is required to update promotion code');
     if (!promoCodeId)
-      throw new Error('promoCodeId is required to updae promotion code');
+      throw new Error('promoCodeId is required to update promotion code');
+    if (!isAffilateInPromoCodeId({ affiliateId, promoCodeId }))
+      throw new Error('Invalid promo code for affiliate');
+
     const cleanPromoCode = scrubPromoCode({ promoCode: upPromoCode });
-    return affiliateCollection
-      .doc(affiliateId)
-      .collection(subCollectionName)
-      .doc(promoCodeId)
+    const affiliatePromoCodeId = promoCodeId;
+    return affiliatePromoCodeCollection
+      .doc(affiliatePromoCodeId)
       .update(cleanPromoCode)
-      .then(() => get({ affiliateId, promoCodeId }));
+      .then(() => get(affiliatePromoCodeId));
   }
 
   function remove({ affiliateId, promoCodeId }) {
     dlog('delete promoCode %s under affiliate %s', promoCodeId, affiliateId);
-    return affiliateCollection
-      .doc(affiliateId)
-      .collection(subCollectionName)
+    if (!isAffilateInPromoCodeId({ affiliateId, promoCodeId }))
+      throw new Error('Invalid promo code for affiliate');
+
+    return affiliatePromoCodeCollection
       .doc(promoCodeId)
       .delete()
       .then(() => ({ id: promoCodeId }));
@@ -133,6 +193,7 @@ const affiliatePromoCode = dbInstance => {
     get,
     getAllAffiliatePromoCodes,
     findAffiliatePromoCodeForEvent,
+    findAllPromoCodesForEvent,
     create,
     update,
     remove,
